Don't crash when Ipsilon has no userinfo
Fixes: #4569
