+++ news/4569.bug
@@ -1,0 +1,1 @@
+Don't crash when Ipsilon has no userinfo
